update auth and lib-auth to reflect api change
