+++ libs/auth/src/index.ts
@@ -2,5 +2,4 @@ export { changePassword } from "./changePassword";
 export { createUser } from "./createUser";
 export type { Capabilities } from "./getCapabilities";
 export { getCapabilities } from "./getCapabilities";
-export { getUser } from "./getUser";
 export { validateToken } from "./validateToken";
